resolved the comments on the PR
