+++ packages/plugin-gtk/src/constants/index.ts
@@ -1,14 +1,6 @@
-import { z } from 'zod';
-
 // Service name constant 
 export const GTK_SERVICE_NAME = 'gtk';
 
-// Configuration schema
-export const configSchema = z.object({
-  API_KEY: z.string().min(1, 'API key is required'),
-  MNEMONIC: z.string().min(1, 'Mnemonic is required'),
-  NETWORK: z.enum(['mainnet', 'testnet']).default('mainnet'),
-});
 
 // Default values
 export const DEFAULT_COLLATERAL_TYPE = 'uusdc';
